feat: rename watchlist to user data module
