+++ app/src/screenshotTest/kotlin/com/divinelink/scenepeek/feature/watchlist/UserDataContentScreenshots.kt
@@ -1,0 +1,11 @@
+package com.divinelink.scenepeek.feature.watchlist
+
+import androidx.compose.runtime.Composable
+import com.divinelink.core.ui.Previews
+import com.divinelink.feature.user.data.UserDataContentPreview
+
+@Composable
+@Previews
+fun UserDataContentScreenshots() {
+  UserDataContentPreview()
+}
