+++ core/model/src/main/kotlin/com/divinelink/core/model/user/data/UserDataParameters.kt
@@ -1,0 +1,9 @@
+package com.divinelink.core.model.user.data
+
+import com.divinelink.core.model.media.MediaType
+
+data class UserDataParameters(
+  val page: Int,
+  val sortBy: UserDataSorting = UserDataSorting.DESCENDING,
+  val mediaType: MediaType,
+)
