+++ core/model/src/main/kotlin/com/divinelink/core/model/user/data/UserDataSorting.kt
@@ -1,0 +1,6 @@
+package com.divinelink.core.model.user.data
+
+enum class UserDataSorting(val value: String) {
+  DESCENDING("desc"),
+  ASCENDING("asc"),
+}
