--- core/model/src/main/kotlin/com/divinelink/core/model/watchlist/WatchlistSorting.kt
@@ -1,6 +1,0 @@
-package com.divinelink.core.model.watchlist
-
-enum class WatchlistSorting(val value: String) {
-  DESCENDING("desc"),
-  ASCENDING("asc"),
-}
