+++ core/scaffold/src/main/kotlin/com/divinelink/core/navigation/NavigationQualifier.kt
@@ -13,7 +13,7 @@ enum class NavigationQualifier {
   LinkHandlingSettings,
   AboutSettings,
   Credits,
-  Watchlist,
+  UserData,
   Onboarding,
   TMDBAuth,
   Profile,
