+++ core/scaffold/src/main/kotlin/com/divinelink/core/navigation/route/UserDataRoute.kt
@@ -1,0 +1,9 @@
+package com.divinelink.core.navigation.route
+
+import androidx.navigation.NavController
+import kotlinx.serialization.Serializable
+
+@Serializable
+object UserDataRoute
+
+fun NavController.navigateToUserData() = navigate(route = UserDataRoute)
