--- core/scaffold/src/main/kotlin/com/divinelink/core/navigation/route/WatchlistRoute.kt
@@ -1,9 +1,0 @@
-package com.divinelink.core.navigation.route
-
-import androidx.navigation.NavController
-import kotlinx.serialization.Serializable
-
-@Serializable
-object WatchlistRoute
-
-fun NavController.navigateToWatchlist() = navigate(route = WatchlistRoute)
