+++ feature/user-data/src/main/res/values/strings.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+  <string name="feature_user_data_movie_tab">Movies</string>
+  <string name="feature_user_data_tv_show_tab">TV Shows</string>
+
+  <string name="feature_user_data_total_movies_in_watchlist">You have %1$d movies in your watchlist</string>
+  <string name="feature_user_data_total_tv_shows_in_watchlist">You have %1$d shows in your watchlist</string>
+
+  <string name="feature_user_data_empty_movies_watchlist">You haven\'t added any movies to your watchlist.</string>
+  <string name="feature_user_data_empty_tv_shows_watchlist">You haven\'t added any TV shows to your watchlist.</string>
+
+  <string name="feature_user_data_login_title">Login to your account</string>
+  <string name="feature_user_data_login_description">You need to be logged in to view your watchlist.</string>
+</resources>
