+++ feature/user-data/src/test/kotlin/com/divinelink/feature/user/data/UserDataViewModelTest.kt
@@ -1,0 +1,354 @@
+package com.divinelink.feature.user.data
+
+import com.divinelink.core.data.session.model.SessionException
+import com.divinelink.core.model.media.MediaItem
+import com.divinelink.core.model.media.MediaType
+import com.divinelink.core.testing.MainDispatcherRule
+import com.divinelink.core.testing.assertUiState
+import com.divinelink.core.testing.expectUiStates
+import com.divinelink.core.testing.factories.model.watchlist.WatchlistResponseFactory
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.test.runTest
+import org.junit.Rule
+import kotlin.test.Test
+
+class UserDataViewModelTest {
+
+  private val testRobot = UserDataViewModelTestRobot()
+
+  @get:Rule
+  val mainDispatcherRule = MainDispatcherRule()
+
+  @Test
+  fun `given user is unauthenticated when view model is created then show unauthenticated error`() =
+    runTest {
+      testRobot
+        .mockObserveAccount {
+          mockFailure(SessionException.Unauthenticated())
+        }
+        .buildViewModel()
+        .assertUiState(
+          expectedUiState = createUiState(
+            forms = mapOf(
+              MediaType.MOVIE to UserDataForm.Error.Unauthenticated,
+              MediaType.TV to UserDataForm.Error.Unauthenticated,
+            ),
+          ),
+        )
+    }
+
+  @Test
+  fun `given generic error when view model is created then show generic error`() = runTest {
+    testRobot
+      .mockObserveAccount {
+        mockFailure(Exception())
+      }
+      .buildViewModel()
+      .assertUiState(
+        expectedUiState = createUiState(
+          forms = mapOf(
+            MediaType.MOVIE to UserDataForm.Error.Unknown,
+            MediaType.TV to UserDataForm.Error.Unknown,
+          ),
+        ),
+      )
+  }
+
+  @Test
+  fun `given user is authenticated when view model is created then fetch watchlist`() = runTest {
+    testRobot
+      .mockObserveAccount {
+        mockSuccess(Result.success(true))
+      }
+      .mockFetchWatchlist {
+        mockSuccess(
+          flowOf(
+            Result.success(WatchlistResponseFactory.movies()),
+            Result.success(WatchlistResponseFactory.tv()),
+          ),
+        )
+      }
+      .buildViewModel()
+      .assertUiState(
+        expectedUiState = createUiState(
+          forms = mapOf(
+            MediaType.MOVIE to UserDataForm.Data(
+              mediaType = MediaType.MOVIE,
+              data = (1..2).map {
+                WatchlistResponseFactory.movies()
+              }.flatMap { it.data },
+              totalResults = WatchlistResponseFactory.movies().totalResults,
+            ),
+            MediaType.TV to UserDataForm.Data(
+              mediaType = MediaType.TV,
+              data = (1..2).map {
+                WatchlistResponseFactory.tv()
+              }.flatMap { it.data },
+              totalResults = WatchlistResponseFactory.tv().totalResults,
+            ),
+          ),
+          pages = mapOf(
+            MediaType.MOVIE to 3,
+            MediaType.TV to 3,
+          ),
+        ),
+      )
+  }
+
+  @Test
+  fun `test onSelectTab along with load more`() = runTest {
+    val dataState = createUiState(
+      forms = mapOf(
+        MediaType.MOVIE to UserDataForm.Data(
+          mediaType = MediaType.MOVIE,
+          data = WatchlistResponseFactory.movies().data + WatchlistResponseFactory.movies().data,
+          totalResults = WatchlistResponseFactory.movies().totalResults,
+        ),
+        MediaType.TV to UserDataForm.Data(
+          mediaType = MediaType.TV,
+          data = WatchlistResponseFactory.tv().data + WatchlistResponseFactory.tv().data,
+          totalResults = WatchlistResponseFactory.tv().totalResults,
+        ),
+      ),
+      pages = mapOf(
+        MediaType.MOVIE to 3,
+        MediaType.TV to 3,
+      ),
+    )
+
+    val updatedIndexState = dataState.copy(selectedTabIndex = 1)
+
+    val updatedTvPageState = updatedIndexState.copy(
+      forms = mapOf(
+        MediaType.MOVIE to UserDataForm.Data(
+          mediaType = MediaType.MOVIE,
+          data = (1..2).map {
+            WatchlistResponseFactory.movies()
+          }.flatMap { it.data },
+          totalResults = WatchlistResponseFactory.movies().totalResults,
+        ),
+        MediaType.TV to UserDataForm.Data(
+          mediaType = MediaType.TV,
+          data = (1..3).map {
+            WatchlistResponseFactory.tv()
+          }.flatMap { it.data },
+          totalResults = WatchlistResponseFactory.tv().totalResults,
+        ),
+      ),
+      pages = mapOf(
+        MediaType.MOVIE to 3,
+        MediaType.TV to 4,
+      ),
+    )
+
+    val expectedUiStates = listOf(
+      dataState,
+      updatedIndexState,
+      updatedTvPageState,
+    )
+
+    testRobot
+      .mockObserveAccount {
+        mockSuccess(Result.success(true))
+      }
+      .mockFetchWatchlist {
+        mockSuccess(
+          flowOf(
+            Result.success(WatchlistResponseFactory.movies()),
+            Result.success(WatchlistResponseFactory.tv()),
+          ),
+        )
+      }
+      .buildViewModel()
+      .expectUiStates(
+        action = {
+          selectTab(1)
+          mockFetchWatchlist {
+            mockSuccess(Result.success(WatchlistResponseFactory.tv()))
+          }
+          onLoadMore()
+        },
+        uiStates = expectedUiStates,
+      )
+  }
+
+  @Test
+  fun `test fetchWatchlist with generic error`() = runTest {
+    testRobot
+      .mockObserveAccount {
+        mockSuccess(Result.success(true))
+      }
+      .buildViewModel()
+      .assertUiState(
+        expectedUiState = createUiState(
+          forms = mapOf(
+            MediaType.MOVIE to UserDataForm.Error.Unknown,
+            MediaType.TV to UserDataForm.Error.Unknown,
+          ),
+        ),
+      )
+  }
+
+  @Test
+  fun `test onLoadMore`() = runTest {
+    testRobot
+      .mockObserveAccount { mockSuccess(Result.success(true)) }
+      .mockFetchWatchlist { mockSuccess(Result.success(WatchlistResponseFactory.movies())) }
+      .buildViewModel()
+      .expectUiStates(
+        action = {
+          mockFetchWatchlist {
+            mockSuccess(
+              Result.success(
+                WatchlistResponseFactory.movies(
+                  page = 2,
+                  canFetchMore = false,
+                ),
+              ),
+            )
+          }
+          onLoadMore()
+        },
+        uiStates = listOf(
+          createUiState(
+            forms = mapOf(
+              MediaType.MOVIE to UserDataForm.Data(
+                mediaType = MediaType.MOVIE,
+                data = WatchlistResponseFactory.movies().data +
+                  WatchlistResponseFactory.movies().data,
+                totalResults = WatchlistResponseFactory.movies().totalResults,
+              ),
+              MediaType.TV to UserDataForm.Loading,
+            ),
+            pages = mapOf(
+              MediaType.MOVIE to 3,
+              MediaType.TV to 1,
+            ),
+          ),
+          createUiState(
+            forms = mapOf(
+              MediaType.MOVIE to UserDataForm.Data(
+                mediaType = MediaType.MOVIE,
+                data = WatchlistResponseFactory.movies().data +
+                  WatchlistResponseFactory.movies().data +
+                  WatchlistResponseFactory.movies(2).data,
+                totalResults = WatchlistResponseFactory.movies().totalResults,
+              ),
+              MediaType.TV to UserDataForm.Loading,
+            ),
+            pages = mapOf(
+              MediaType.MOVIE to 4,
+              MediaType.TV to 1,
+            ),
+            canFetchMore = mapOf(
+              MediaType.MOVIE to false,
+              MediaType.TV to true,
+            ),
+          ),
+        ),
+      )
+  }
+
+  @Test
+  fun `test onRefresh when state is unknown error for movie`() = runTest {
+    testRobot
+      .mockObserveAccount {
+        mockFailure(Exception())
+      }
+      .buildViewModel()
+      .assertUiState(
+        expectedUiState = createUiState(
+          forms = mapOf(
+            MediaType.MOVIE to UserDataForm.Error.Unknown,
+            MediaType.TV to UserDataForm.Error.Unknown,
+          ),
+        ),
+      )
+      .mockFetchWatchlist {
+        mockSuccess(
+          flowOf(Result.success(WatchlistResponseFactory.movies())),
+        )
+      }
+      .onRefresh()
+      .assertUiState(
+        expectedUiState = createUiState(
+          forms = mapOf(
+            MediaType.MOVIE to UserDataForm.Data(
+              mediaType = MediaType.MOVIE,
+              data = WatchlistResponseFactory.movies().data,
+              totalResults = WatchlistResponseFactory.movies().totalResults,
+            ),
+            MediaType.TV to UserDataForm.Error.Unknown,
+          ),
+          pages = mapOf(
+            MediaType.MOVIE to 2,
+            MediaType.TV to 1,
+          ),
+        ),
+      )
+  }
+
+  @Test
+  fun `test onRefresh when state is unknown error for tv`() = runTest {
+    testRobot
+      .mockObserveAccount {
+        mockFailure(Exception())
+      }
+      .buildViewModel()
+      .assertUiState(
+        expectedUiState = createUiState(
+          forms = mapOf(
+            MediaType.MOVIE to UserDataForm.Error.Unknown,
+            MediaType.TV to UserDataForm.Error.Unknown,
+          ),
+        ),
+      )
+      .selectTab(1)
+      .mockFetchWatchlist {
+        mockSuccess(
+          flowOf(Result.success(WatchlistResponseFactory.tv())),
+        )
+      }
+      .onRefresh()
+      .assertUiState(
+        expectedUiState = createUiState(
+          forms = mapOf(
+            MediaType.MOVIE to UserDataForm.Error.Unknown,
+            MediaType.TV to UserDataForm.Data(
+              mediaType = MediaType.TV,
+              data = WatchlistResponseFactory.tv().data,
+              totalResults = WatchlistResponseFactory.tv().totalResults,
+            ),
+          ),
+          selectedTabIndex = 1,
+          pages = mapOf(
+            MediaType.MOVIE to 1,
+            MediaType.TV to 2,
+          ),
+        ),
+      )
+  }
+
+  private fun createUiState(
+    selectedTabIndex: Int = 0,
+    tabs: List<MediaTab> = listOf(MediaTab.MOVIE, MediaTab.TV),
+    pages: Map<MediaType, Int> = mapOf(
+      MediaType.MOVIE to 1,
+      MediaType.TV to 1,
+    ),
+    forms: Map<MediaType, UserDataForm<MediaItem.Media>> = mapOf(
+      MediaType.MOVIE to UserDataForm.Loading,
+      MediaType.TV to UserDataForm.Loading,
+    ),
+    canFetchMore: Map<MediaType, Boolean> = mapOf(
+      MediaType.MOVIE to true,
+      MediaType.TV to true,
+    ),
+  ): UserDataUiState = UserDataUiState(
+    selectedTabIndex = selectedTabIndex,
+    tabs = tabs,
+    pages = pages,
+    forms = forms,
+    canFetchMore = canFetchMore,
+  )
+}
