+++ settings.gradle.kts
@@ -37,8 +37,8 @@ include(":core:fixtures")
 include(":feature:details")
 include(":feature:search")
 include(":feature:settings")
-include(":feature:watchlist")
 include(":feature:credits")
 include(":feature:tmdb-auth")
 include(":feature:onboarding")
 include(":feature:profile")
+include(":feature:user-data")
